clean up types and fix types issues
